Refactor reward distributions & reward distributor IT
