+++ mobile_verifier/tests/integrations/main.rs
@@ -7,6 +7,7 @@ mod heartbeats;
 mod hex_boosting;
 mod last_location;
 mod modeled_coverage;
+mod reward_distributor;
 mod rewarder_poc_dc;
 mod rewarder_sp_rewards;
 mod seniority;
